refactor toPath/fromPath
The signatures of toPath and fromPath are changed such that instead of
receiving the raw input value and the URL encoded/decoded value as
appropriate, they receive only the raw value (i.e. what was matched in
the URL, or what the user passed in the redux action). Responsibility
for URL encoding and decoding now lies completely in toPath/fromPath.

In doing so a bug was fixed where null params were serialised as the
string "null". null params are now treated as though they were not
provided, although this behaviour can be overriden in toPath.