Handle URL encoding and decoding internally
Previously fromPath and toPath were responsible for doing URL encoding
and decoding. Now path-to-regexp's built in URL encoding is used when
creating URLs, and we consistently decode segments on the way out
(separately from toPath and fromPath).

toPath and fromPath have been refactored such that for multi segment
params, the URL side is an array of segments. instead of a string
containing separators. This means that encoding is complete, i.e. it is
possible to put any character you want in params, including slashes.
toPath and fromPath also receive the `optional` and `repeat` keys from
path-to-regexp so that they can determine whether to use arrays or
strings.

The default behaviour is similar to before, with strings for multi
segment params, but toPath throws for unrecognised param types, and
exceptions thrown in toPath result in 404s.