feat: add compatibility with react-redux 6+ (#33)
The method of making the Rudy API available from within the link component had to change since it relied on the old React context used by react-redux 5 and below. In the newly renamed `react` package there is a `RudyProvider` and a `RudyConsumer` which internally use the new React context API to pass the Rudy API through the component tree.

BREAKING CHANGE: The object returned from createRouter now contains a key `api` which contains `routes`, `ctx` and such things. Previously they were in the root object.

BREAKING CHANGE: The `link` package has been renamed to `react`.

BREAKING CHANGE: The `Link` component now must be rendered as a descendant of the new `RudyProvider`.

BREAKING CHANGE: The core no longer monkey patches redux `getState` to have a key `api` with the rudy API.