feat: fallback for missing sessionStorage
Although all modern browsers support sessionStorage, there are various
circumstances in which it fails to work:
- if the user fills the entire quota
- in private browsing mode on iOS Safari
- if all cookies are disabled in Chrome

This change implements a fallback in this case to avoid errors. Instead
of pushState, history pushes result in a full page navigation. This
ensures that the length of the history that we need to mirror is always
1, and so we can fallback to JS module state for sessionStorage, which
always works.